Move useProperty hook to hooks folder
