+++ src/components/Property/Types/ListProperty/IntListProperty.tsx
@@ -1,6 +1,6 @@
 import { Pills } from '@/components/Pills/Pills';
 import { PropertyProps } from '@/components/Property/types';
-import { useProperty } from '@/types/hooks';
+import { useProperty } from '@/hooks/properties';
 
 export function IntListProperty({ uri, readOnly }: PropertyProps) {
   const [value, setValue] = useProperty('IntListProperty', uri);
