+++ src/components/Property/Types/StringProperty.tsx
@@ -2,7 +2,7 @@ import { Paper, Text } from '@mantine/core';
 
 import { StringInput } from '@/components/Input/StringInput';
 import { PropertyProps } from '@/components/Property/types';
-import { useProperty } from '@/types/hooks';
+import { useProperty } from '@/hooks/properties';
 
 export function StringProperty({ uri, readOnly }: PropertyProps) {
   const [value, setValue, meta] = useProperty('StringProperty', uri);
