+++ src/panels/GettingStartedPanel/Tasks/AltitudeTask.tsx
@@ -4,7 +4,7 @@ import { NavigationAnchorKey } from '@/util/keys';
 
 import { TaskCheckbox } from './Components/TaskCheckbox';
 import { compareAltitude } from './util';
-import { useProperty } from '@/types/hooks';
+import { useProperty } from '@/hooks/properties';
 
 interface Props {
   anchor: string;
