+++ src/panels/GettingStartedPanel/Tasks/Components/CurrentFocus.tsx
@@ -1,7 +1,7 @@
 import { Text } from '@mantine/core';
 
 import { NavigationAnchorKey } from '@/util/keys';
-import { useProperty } from '@/types/hooks';
+import { useProperty } from '@/hooks/properties';
 
 export function CurrentFocus() {
   const [currentAnchor] = useProperty('StringProperty', NavigationAnchorKey);
