+++ src/panels/GettingStartedPanel/Tasks/FocusTask.tsx
@@ -1,7 +1,7 @@
 import { NavigationAnchorKey } from '@/util/keys';
 
 import { TaskCheckbox } from './Components/TaskCheckbox';
-import { useProperty } from '@/types/hooks';
+import { useProperty } from '@/hooks/properties';
 
 interface Props {
   anchor: string;
